requirements.txt
Flask>=2.3.0,<3.0.0
firebase-admin>=6.4.0,<7.0.0
PyJWT>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
Werkzeug>=2.3.0,<3.0.0